+++ NEWS.md
@@ -1,6 +1,8 @@
 # longears 0.1.0.9000
 
 - Tests requiring a local RMQ server will now be skipped unless one is detected.
+- Fixes an issue where `amqp_properties()` would not handle missing names
+  correctly.
 
 # longears 0.1.0 (2019-06-06)